+++ modules/pathkit/pathkit_wasm_bindings.cpp
@@ -393,6 +393,7 @@ struct StrokeOpts {
     // any are omitted.
     SkScalar width;
     SkScalar miter_limit;
+    SkScalar res_scale;
     SkPaint::Join join;
     SkPaint::Cap cap;
 };
@@ -404,8 +405,11 @@ bool ApplyStroke(SkPath& path, StrokeOpts opts) {
     p.setStrokeJoin(opts.join);
     p.setStrokeWidth(opts.width);
     p.setStrokeMiter(opts.miter_limit);
-
-    return p.getFillPath(path, &path);
+    // Default to 1.0 if 0 (or an invalid negative number)
+    if (opts.res_scale <= 0) {
+        opts.res_scale = 1.0;
+    }
+    return p.getFillPath(path, &path, nullptr, opts.res_scale);
 }
 
 //========================================================================================
@@ -589,6 +593,7 @@ EMSCRIPTEN_BINDINGS(skia) {
     value_object<StrokeOpts>("StrokeOpts")
         .field("width",       &StrokeOpts::width)
         .field("miter_limit", &StrokeOpts::miter_limit)
+        .field("res_scale",   &StrokeOpts::res_scale)
         .field("join",        &StrokeOpts::join)
         .field("cap",         &StrokeOpts::cap);