+++ modules/pathkit/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - `res_scale` expoed in StrokeOpts object. (skbug.com/13301)
+
 ## [1.0.0] 2022-02-03
 
 ### Changed